+++ realwork/realwork/ArtistMediaTableViewController..swift
@@ -1,0 +1,9 @@
+//
+//  ArtistMediaTableViewController..swift
+//  realwork
+//
+//  Created by Nathan Ngobale on 2020/02/19.
+//  Copyright © 2020 Nathan Ngobale. All rights reserved.
+//
+
+import Foundation
