+++ realwork/realwork/ArtistMediaTableViewController.swift
@@ -1,0 +1,105 @@
+//
+//  ArtistMediaTableViewController.swift
+//  realwork
+//
+//  Created by Nathan Ngobale on 2020/02/19.
+//  Copyright © 2020 Nathan Ngobale. All rights reserved.
+//
+
+import UIKit
+
+class ArtistMediaTableViewController: UITableViewController {
+    var listOfArtistCollections = [Collection]() {
+        didSet {
+//            print(self.listOfArtistCollections)
+//            print("XXXXXXXXX")
+            DispatchQueue.main.async {
+                self.tableView.reloadData()
+            }
+        }
+    }
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Uncomment the following line to preserve selection between presentations
+        // self.clearsSelectionOnViewWillAppear = false
+        // Uncomment the following line to display an Edit button in the navigation bar for this view controller.
+        // self.navigationItem.rightBarButtonItem = self.editButtonItem
+        getInformation()
+//        print(listOfArtistCollections)
+    }
+    // MARK: - Table view data source
+
+
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        // #warning Incomplete implementation, return the number of rows
+        return listOfArtistCollections.count
+    }
+
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: "Cell", for: indexPath)
+
+        // Configure the cell...
+        cell.textLabel?.text = listOfArtistCollections[indexPath.row].artistName
+        return cell
+    }
+
+    /*
+     // Override to support conditional editing of the table view.
+     override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
+     // Return false if you do not want the specified item to be editable.
+     return true
+     }
+     */
+
+    /*
+     // Override to support editing the table view.
+     override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
+     if editingStyle == .delete {
+     // Delete the row from the data source
+     tableView.deleteRows(at: [indexPath], with: .fade)
+     } else if editingStyle == .insert {
+     // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
+     }
+     }
+     */
+    
+    /*
+     // Override to support rearranging the table view.
+     override func tableView(_ tableView: UITableView, moveRowAt fromIndexPath: IndexPath, to: IndexPath) {
+     
+     }
+     */
+    
+    /*
+     // Override to support conditional rearranging of the table view.
+     override func tableView(_ tableView: UITableView, canMoveRowAt indexPath: IndexPath) -> Bool {
+     // Return false if you do not want the item to be re-orderable.
+     return true
+     }
+     */
+    
+    /*
+     // MARK: - Navigation
+     
+     // In a storyboard-based application, you will often want to do a little preparation before navigation
+     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+     // Get the new view controller using segue.destination.
+     // Pass the selected object to the new view controller.
+     }
+     */
+    
+    func getInformation() {
+        let name = "beyonce"
+        let artistMediaRequest = ArtistMediaRequest(artistName: name)
+        artistMediaRequest.getArtistMedia { [weak self] result in
+            switch result {
+            case .failure(let error):
+                print(error)
+            case .success(let collections):
+                self?.listOfArtistCollections = collections
+            }
+        }
+    }
+    
+}
